ros2_ws/src/drone_tracker/requirements.txt
ament_copyright==0.12.11
ament_flake8==0.12.11
ament_index_python==1.4.0
ament_pep257==0.12.11
cv_bridge==3.2.1
keyboard==0.13.5
launch_ros==0.19.7
numpy==1.21.5
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pyserial==3.5
pytest==6.2.5
rcl_interfaces==1.2.1
rclpy==3.3.13
sensor_msgs==4.2.4
setuptools==59.6.0
std_msgs==4.2.4
torch==2.3.0
ultralytics==8.2.98
